backend/requirements.txt
bcrypt==5.0.0
email-validator
fastapi
google-generativeai
passlib[bcrypt]
psycopg2-binary
pydantic-settings
python-jose
python-multipart
python-dotenv
SQLAlchemy
uvicorn
cloudinary
PyMuPDF
